udpated i3 dependencies and setup

+++ install-scripts/basic/i3.sh
@@ -1,17 +1,32 @@
-sudo apt install jq i3 feh nitrogen compton terminology zenity rofi python3
+sudo apt install jq i3 feh nitrogen compton terminology zenity rofi python3 python3-pip curl libxcb1-dev udiskie  volumeicon-alsa
 
-# installing this for having graphical interface when setting/going to marks
-pip3 install git+https://github.com/talwrii/i3-rofi-mark#egg=i3-rofi-mark
+# install rust/cargo if not alreay installed
+if ! [ -x "$(command -v cargo)" ]; then
+	curl --proto '=https' --tlsv1.2 -sSf https://sh.rustup.rs | sh
+	echo "Need to reboot before proceeding..."
+	echo "Reboot and execute this script again"
+else
 
-# installing i3wsr
-cargo install i3wsr
+	# install rust
+	curl --proto '=https' --tlsv1.2 -sSf https://sh.rustup.rs | sh
 
-# i3
-echo "Configuring i3"
-if [ -e ~/.i3 ]; then
-    echo "Deleting i3 directory"
-    rm -rf ~/.i3
+	# installing i3wsr
+	cargo install i3wsr
+
+	# installing oh-my-zsh
+	source ./zsh.sh
+
+	# i3
+	echo "Configuring i3"
+	if [ -e ~/.i3 ]; then
+		echo "Deleting i3 directory"
+		rm -rf ~/.i3
+	fi
+	if [ -e $dotfilesdir/config-files/i3 ];then
+		if [ -e $USER/.i3 ]; then
+			rm -rf $USER/.i3
+		fi
+		ln -s $dotfilesdir/config-files/.i3 /home/$USER/.i3
+	fi
 fi
-if [ -e $dotfilesdir/config-files/i3 ];then
-    ln -s $dotfilesdir/config-files/.i3 /home/$USER/.i3
-fi
+
